# chat-backend/docs/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "TAI后端服务",
    "title": "TAI Backend API",
    "version": "1.0.0"
  },
  "host": "localhost:9090",
  "basePath": "/api/v1",
  "paths": {
...
    "/models": {
      "post": {
        "description": "保存或更新模型的配置信息，包括API密钥、端点等设置",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Models"
        ],
        "summary": "保存模型配置",
        "operationId": "saveModel",
        "parameters": [
          {
            "description": "模型信息",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "description": "模型信息",
              "type": "object",
              "$ref": "#/definitions/ModelSaveRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ModelID",
            "schema": {
              "$ref": "#/definitions/ModelID"
            }
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/models/available/all": {
      "get": {
        "description": "获取当前系统中所有可用的模型列表，包括聊天模型和向量模型",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Models"
        ],
        "summary": "获取所有可用模型",
        "operationId": "listAvailableAllModels",
        "responses": {
          "200": {
            "$ref": "#/responses/models.ModelInfo"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/models/available/chat": {
      "get": {
        "description": "获取当前系统中可用的聊天模型列表，仅包含已配置且可用的聊天模型",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Models"
        ],
        "summary": "获取可用聊天模型",
        "operationId": "listAvailableChatModels",
        "responses": {
          "200": {
            "$ref": "#/responses/models.ModelInfo"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/models/available/vector": {
      "get": {
        "description": "获取当前系统中可用的向量模型列表，用于知识库文档向量化",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Models"
        ],
        "summary": "获取可用向量模型",
        "operationId": "listAvailableVectorModels",
        "responses": {
          "200": {
            "$ref": "#/responses/models.ModelInfo"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/models/supported/chat": {
      "get": {
        "description": "获取系统支持的所有聊天模型列表，包括模型名称、类型等信息",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Models"
        ],
        "summary": "获取支持的聊天模型",
        "operationId": "listSupportedChatModels",
        "responses": {
          "200": {
            "$ref": "#/responses/models.SupportedChatModel"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/models/supported/vector": {
      "get": {
        "description": "获取系统支持的所有向量模型列表，用于知识库向量化处理",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Models"
        ],
        "summary": "获取支持的向量模型",
        "operationId": "listSupportedVectorModels",
        "responses": {
          "200": {
            "$ref": "#/responses/models.SupportedVectorModel"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/models/{id}": {
      "delete": {
        "description": "根据模型ID删除指定的模型配置信息",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Models"
        ],
        "summary": "删除模型配置",
        "operationId": "deleteModel",
        "parameters": [
          {
            "type": "integer",
            "description": "模型ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/MessageOnlyResponse"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
    "/models/{id}/status": {
      "put": {
        "description": "启用或禁用指定的模型，控制模型是否可用于对话",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Models"
        ],
        "summary": "设置模型状态",
        "operationId": "setModelStatus",
        "parameters": [
          {
            "type": "integer",
            "description": "模型ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "状态设置",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "description": "状态设置",
              "type": "object",
              "$ref": "#/definitions/ModelStatusEnableRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/MessageOnlyResponse"
          },
          "400": {
            "description": "ResponseBody",
            "schema": {
              "$ref": "#/definitions/ResponseBody"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ModelID": {
      "description": "ModelID 模型ID响应数据",
      "type": "object",
      "required": [
        "id"
      ],
      "properties": {
        "id": {
          "description": "模型ID",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        }
      },
      "x-go-package": "chat-backend/models"
    },
...
    "ModelSaveRequest": {
      "description": "ModelSaveRequest 添加/修改模型请求\nAPI: POST /model/save",
      "type": "object",
      "required": [
        "id",
        "name",
        "type",
        "symbol",
        "endpoint",
        "enable",
        "credentials"
      ],
      "properties": {
        "credentials": {
          "description": "模型鉴权信息",
          "type": "string",
          "x-go-name": "Credentials"
        },
        "enable": {
          "description": "是否启用",
          "type": "boolean",
          "x-go-name": "Enable"
        },
        "endpoint": {
          "description": "模型接入点",
          "type": "string",
          "x-go-name": "Endpoint"
        },
        "id": {
          "description": "模型ID: 0=新建, \u003e0=修改",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ID"
        },
        "name": {
          "description": "模型名称",
          "type": "string",
          "x-go-name": "Name"
        },
        "symbol": {
          "description": "模型编码",
          "type": "string",
          "x-go-name": "Symbol"
        },
        "type": {
          "description": "模型类型: 0=聊天模型, 1=嵌入模型",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Type"
        }
      },
      "x-go-package": "flowy-sdk/services/model"
    },
    "ModelStatusEnableRequest": {
      "description": "ModelStatusEnableRequest 模型状态启用/禁用请求",
      "type": "object",
      "required": [
        "enable"
      ],
      "properties": {
        "enable": {
          "description": "true=启用, false=禁用",
          "type": "boolean",
          "x-go-name": "Enable"
        }
      },
      "x-go-package": "chat-backend/models"
    },
...
    "ResponseBody": {
      "description": "ResponseBody 统一响应体结构",
      "type": "object",
      "required": [
        "is_login",
        "code",
        "msg",
        "data",
        "time",
        "code_num",
        "code_msg"
      ],
      "properties": {
        "code": {
          "description": "请求是否成功",
          "type": "boolean",
          "x-go-name": "Code"
        },
        "code_msg": {
          "description": "错误代码消息",
          "type": "string",
          "x-go-name": "CodeMsg"
        },
        "code_num": {
          "description": "数字错误代码",
          "type": "integer",
          "format": "int64",
          "x-go-name": "CodeNum"
        },
        "data": {
          "description": "响应数据",
          "x-go-name": "Data"
        },
        "is_login": {
          "description": "登录状态",
          "type": "boolean",
          "x-go-name": "IsLogin"
        },
        "msg": {
          "description": "响应消息",
          "type": "string",
          "x-go-name": "Message"
        },
        "time": {
          "description": "持续时间（秒）",
          "type": "number",
          "format": "double",
          "x-go-name": "Time"
        }
      },
      "x-go-package": "chat-backend/utils"
    },
...
  },
  "responses": {
...
    "MessageOnlyResponse": {
      "description": "MessageOnlyResponse 仅消息成功响应",
      "schema": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "description": "响应消息",
            "type": "string",
            "x-go-name": "Message"
          }
        }
      }
    },
...
  }
}